Code clean up and fix starboard

+++ Yami/plugins/events.py
@@ -7,7 +7,7 @@ from hikari import (
     ShardReadyEvent,
     GuildAvailableEvent,
     ReactionDeleteEvent,
-    ReactionAddEvent, NotFound,
+    ReactionAddEvent, NotFoundError,
 )
 from lightbulb import plugins
 
@@ -55,43 +55,55 @@ class Events(plugins.Plugin):
     @plugins.listener(event_type=ReactionAddEvent)
     async def on_reaction_add(self, event: ReactionAddEvent):
         guild, _ = await models.Guild.get_or_create(id=event.guild_id)
-        if not guild.starboard:
+        if not guild.starboard or not event.emoji == "⭐":
             return
-        if not event.emoji == "⭐":
-            return
+
         # noinspection PyTypeChecker
         starboard: hikari.GuildTextChannel = await self.bot.rest.fetch_channel(
             guild.starboard
         )
+
         message = await self.bot.rest.fetch_message(event.channel_id, event.message_id)
-        if message.author.id == event.user_id:
+
+        if message.author.id == event.user_id or message.author.id == self.bot.user.id or event.member.is_bot:
             return await message.remove_reaction(emoji="⭐", user=message.author)
-        if message.author.id == self.bot.user.id or event.member.is_bot:
-            return
+
         starred_message_model = await models.StarredMessage.get_or_none(
             id=event.message_id
         )
+
         stars = await self.bot.rest.fetch_reactions_for_emoji(
             event.channel_id, event.message_id, "⭐"
         ).count()
+
         if starred_message_model:
             try:
                 starred_message = await self.bot.rest.fetch_message(
                     starboard, starred_message_model.star_id
                 )
-            except NotFound:
+            except NotFoundError:
                 await starred_message_model.delete()
                 return
+
             embed, content = await self.get_starboard_embed(
                 message, guild.id, starred_message.created_at, stars
             )
+
+            starred_message_model.stars = stars
+            await starred_message_model.save()
+
             await starred_message.edit(embed=embed, content=content)
+
         else:
-            if stars > 2:
+
+            if stars >= 2:
+
                 embed, content = await self.get_starboard_embed(
                     message, guild.id, datetime.now(timezone.utc), stars
                 )
+
                 starred_message = await starboard.send(embed=embed, content=content)
+
                 await models.StarredMessage.create(
                     id=message.id, star_id=starred_message.id, stars=stars
                 )
@@ -99,34 +111,42 @@ class Events(plugins.Plugin):
     @plugins.listener(event_type=ReactionDeleteEvent)
     async def on_reaction_remove(self, event: ReactionDeleteEvent):
         guild, _ = await models.Guild.get_or_create(id=event.guild_id)
-        if not guild.starboard:
+
+        if not guild.starboard or not event.emoji == "⭐":
             return
-        if not event.emoji == "⭐":
-            return
+
         # noinspection PyTypeChecker
         starboard: hikari.GuildTextChannel = await self.bot.rest.fetch_channel(
             guild.starboard
         )
+
         message = await self.bot.rest.fetch_message(event.channel_id, event.message_id)
+
         if event.user_id == message.author.id:
             return
+
         starred_message_model = await models.StarredMessage.get_or_none(
             id=event.message_id
         )
+
         stars = await self.bot.rest.fetch_reactions_for_emoji(
             event.channel_id, event.message_id, "⭐"
         ).count()
-        if starred_message_model:
+
+        if starred_message_model and stars > starred_message_model.stars:
+
             try:
                 starred_message = await self.bot.rest.fetch_message(
                     starboard, starred_message_model.star_id
                 )
-            except NotFound:
+            except NotFoundError:
                 await starred_message_model.delete()
                 return
             embed, content = await self.get_starboard_embed(
                 message, guild.id, starred_message.created_at, stars
             )
+            starred_message_model.stars = stars
+            await starred_message_model.save()
             await starred_message.edit(embed=embed, content=content)
